Move functions to avoid circular dependencies
